Refactor PostsView UI

+++ lib/features/posts/presentation/widgets/top_post_item_section.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:social_app/config/themes/app_colors.dart';
+import 'package:social_app/config/themes/app_text_styles.dart';
 import 'package:social_app/core/helpers/helper.dart';
 import 'package:social_app/core/widgets/post_information.dart';
 import 'package:social_app/features/posts/data/models/post_model.dart';
@@ -20,16 +22,29 @@ class TopPostItemSection extends StatelessWidget {
       children: <Widget>[
         Flexible(child: PostInformation(post: post)),
         post.uId == Helper.currentUser!.uId
-            ? IconButton(
-                onPressed: () {
-                  BlocProvider.of<PostsCubit>(context)
-                      .deletePost(postId: post.postId!);
-                },
+            ? PopupMenuButton(
                 icon: Icon(
-                  Icons.delete,
+                  Icons.more_horiz,
+                  color: Helper.isDark(context)
+                      ? AppColors.scaffoldBackgroundClr
+                      : AppColors.darkPrimaryColor,
                   size: 25.w,
-                  color: Colors.red,
                 ),
+                itemBuilder: (context) => [
+                  PopupMenuItem(
+                    value: 'deletePost',
+                    child: Text(
+                      'Delete',
+                      style: AppTextStyles.textStyle16SemiBold,
+                    ),
+                  ),
+                ],
+                onSelected: (value) async {
+                  if (value == 'deletePost') {
+                    BlocProvider.of<PostsCubit>(context)
+                        .deletePost(postId: post.postId!);
+                  }
+                },
               )
             : const SizedBox(),
       ],
